feat: Add error handling in redis

+++ config/redis.js
@@ -1,8 +1,7 @@
-/* eslint-disable object-shorthand */
-/* eslint-disable func-names */
 const Redis = require('ioredis');
 require('dotenv').config();
 const { CACHE_PORT, CACHE_HOST, CACHE_USER, CACHE_PASSWORD } = process.env;
+
 const redis = new Redis({
     port: CACHE_PORT,
     host: CACHE_HOST,
@@ -14,6 +13,7 @@ const redis = new Redis({
         return delay;
     },
 });
+
 redis.ready = false;
 redis.on('ready', async () => {
     redis.ready = true;
@@ -23,14 +23,27 @@ redis.on('ready', async () => {
 redis.on('error', async () => {
     if (redis.ready) {
         console.error('Error in Redis');
-        console.error('Redis connection error');
+        console.error('Redis lose connection');
     }
     redis.ready = false;
 });
 
-const clearRedisDb = async () => {
-    console.warn('Cleaning Redis');
-    await redis.flushdb();
+const clearRedis = async () => {
+    if (redis.ready) {
+        console.warn('Cleaning Redis!!!');
+        await redis.flushdb();
+    }
 };
-clearRedisDb(); // Clean Redis when an app initializes
+
+const testRedisConnection = async () => {
+    if (!redis.ready) {
+        redis.connect().catch((err) => {
+            console.log(err);
+            console.log('redis connect fail');
+        });
+    }
+};
+
+setTimeout(testRedisConnection, 500);
+setTimeout(clearRedis, 400);
 module.exports = redis;
